nft-extension/backend/requirements.txt
--find-links https://prototyp3-dev.github.io/pip-wheels-riscv/wheels/
base58==2.1.1
cartesapp @ git+https://github.com/prototyp3-dev/cartesapp@84547bb16e012f6a6abcb2d023a1baacb78982e1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cytoolz==0.12.3
eth-abi==4.2.1
eth-hash==0.7.0
eth-typing==4.2.1
eth-utils==4.1.0
exceptiongroup==1.2.1
idna==3.7
iniconfig==2.0.0
markdown-it-py==3.0.0
mdurl==0.1.2
packaging==24.0
parsimonious==0.9.0
pillow==10.2.0
pluggy==1.4.0
pony==0.7.17
protobuf==4.25.3
py-expression-eval==0.3.14
pycryptodome==3.19.1
pydantic==1.10.15
Pygments==2.17.2
pytest==7.4.4
python-cartesi==0.1.1
regex==2024.4.16
requests==2.31.0
rich==13.7.1
shellingham==1.5.4
tomli==2.0.1
toolz==0.12.1
typer==0.12.3
typing_extensions==4.11.0
urllib3==2.2.1
